CONEXCION COMPLETA A API DE RPOYECTOS

+++ src/main/java/com/psa/soporte/DTO/response/TareaResponse.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class TareaResponse {
-    private Long tareaId;
+    private Integer tareaId;
     private String descripcion;
 }
